copy symbols in wrapper_create
We don't know the lifetime of the const char* input, so lets copy it.
this is quite naive, solution for now, but should be good enough as this
is mainly for debugging. Maybe leter do something more fancy to avoid
copies from symbols that came from linker for example, and use hashtable
to avoid duplication.